Passing function as an object, learned about this and its limitations, syntax parsing, power of  arrays
Biggest lesson I learned is how this reference works

function expression or statement created in the global scope, points to global

function or method within an object, points to object.

method that has another method within it, the outer points to the object and the inner points to the window object.